Invalid Player Fixed
a previous glitch would not break out of a loop once the user entered an invalid player code. Now the user has three opportunities to enter a valid player code to continue the game when selecting a shooting guard. this code will now be used for the other four positions.